fix: Resolve MSVC type conversion warnings in rendering files
- Add explicit static_cast<float>(M_PI) in RaytracingPipeline.cpp line 330
- Add explicit static_cast<float>(M_PI) in Light.cpp lines 177-178
- Fixes C4244 warnings for double to float conversion in MSVC builds

+++ src/Rendering/Lighting/Light.cpp
@@ -174,8 +174,8 @@ float Light::GetSpotAttenuation(const Vector3& lightToPoint) const {
     Vector3 pointDirection = lightToPoint.Normalized();
     
     float cosAngle = lightDirection.Dot(pointDirection);
-    float innerCos = std::cos(m_data.innerConeAngle * 0.5f * M_PI / 180.0f);
-    float outerCos = std::cos(m_data.outerConeAngle * 0.5f * M_PI / 180.0f);
+    float innerCos = std::cos(m_data.innerConeAngle * 0.5f * static_cast<float>(M_PI) / 180.0f);
+    float outerCos = std::cos(m_data.outerConeAngle * 0.5f * static_cast<float>(M_PI) / 180.0f);
     
     if (cosAngle > innerCos) {
         return 1.0f; // Full intensity
